Latest work.
We were doing well before but it was kerning *everything* negatively. Since most kern pairs in a font are zero, we want to expose the network to this fact, rather than artificially balancing the classes. We then do a bit of class weighting, and use a custom loss function which penalizes false negatives - if we don’t do that, the network takes the easy option and just guesses 0 all the time.